% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format-results.R
\name{add_sens_FDP}
\alias{add_sens_FDP}
\title{Alternative sensitivity & False Discovery Proportion}
\usage{
add_sens_FDP(x, corr)
}
\arguments{
\item{x}{Data frame containing variables "true_set" and "set".}

\item{corr}{Sparse correlation matrix of the genotype matrix.}
}
\value{
\code{x} with 2 new variables "AltSens" and "AltFDP".
}
\description{
Compute alternative sensitivity & False Discovery Proportion from
https://doi.org/10.1289/EHP172.
}
