% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format-results.R
\name{boot}
\alias{boot}
\title{Bootstrap}
\usage{
boot(x, n = 1e+05, f = mean)
}
\arguments{
\item{x}{Vector of results.}

\item{n}{Number of bootstrap replicate.}

\item{f}{Function to apply to samples of \code{x}. Default is \code{mean}.}
}
\value{
Standard deviation of bootstrap results.
}
\description{
Bootstrap
}
