% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format-results.R
\name{c_method_r2}
\alias{c_method_r2}
\title{Combine "method" and "r2"}
\usage{
c_method_r2(method, r2)
}
\arguments{
\item{method}{Vector of names of methods.}

\item{r2}{Vector of r2 values (possibly with missing values).}
}
\value{
Concatenation of \code{method} and \code{r2} using \code{"-"}.
}
\description{
Combine "method" and "r2"
}
\examples{
c_method_r2(c("A", "B", "A"), c(0.5, NA, 0.2))
}
