% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{get_pheno}
\alias{get_pheno}
\title{Simulate phenotypes}
\usage{
get_pheno(G, h2, M, ind.possible = bigstatsr::cols_along(G),
  effects.dist = c("gaussian", "laplace"), model = c("simple",
  "fancy"), K = 0.3)
}
\arguments{
\item{G}{Matrix of genotypes.}

\item{h2}{Heritability.}

\item{M}{Number of causal variants. Should be multiple of 6.}

\item{ind.possible}{Indices of possible causal variants.}

\item{effects.dist}{Distribution of effects.
Either \code{"gaussian"} (the default) or \code{"laplace"}.}

\item{model}{Model for simulation.
Either \code{"simple"} (the default) with linear effets only,
or \code{"fancy"} with linear, domination and interaction-type effetcs.}

\item{K}{Prevalence. Default is \code{0.3}.}
}
\value{
A list with 3 elements:
\itemize{
\item "pheno": vector of binary phenotypes,
\item "set": indices of causal variants,
\item "effects": Size of effects corresponding to \code{set}.
}
}
\description{
Simulate phenotypes
}
