% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.paper.PRS-package.R
\docType{package}
\name{pkg.paper.PRS-package}
\alias{pkg.paper.PRS}
\alias{pkg.paper.PRS-package}
\title{pkg.paper.PRS: Functions Used in Paper 2}
\description{
Functions that are used in my second paper (about PRS).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
