% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{PRS}
\alias{PRS}
\title{PRS (Clumping + Thresholding)}
\usage{
PRS(G, infos.chr, infos.pos, pheno.all, covar.all, ind.train, ind.test,
  ncores = nb_cores(), family = "logistic")
}
\arguments{
\item{G}{A \link[=FBM.code256-class]{FBM.code256}
(typically \code{<bigSNP>$genotypes}).\cr
\strong{You shouldn't have missing values in your data or SNPs with 0 MAF.}}

\item{infos.chr}{Vector of integers specifying each SNP's chromosome.\cr
Typically \code{<bigSNP>$map$chromosome}.}

\item{infos.pos}{Vector of integers specifying the physical position
on a chromosome (in base pairs) of each SNP.\cr
Typically \code{<bigSNP>$map$physical.pos}.}

\item{pheno.all}{All phenotypes.}

\item{covar.all}{All covariables.}

\item{ind.train}{Indices corresponding to the training set.}

\item{ind.test}{Indices corresponding to the test set.}

\item{ncores}{Number of cores to use. Default is \code{bigstatsr::nb_cores()}.}

\item{family}{Default is \code{"logistic"}. Can also use \code{"linear"}.}
}
\value{
A data frame with 5 variables:
\itemize{
\item "method": the name of the method,
\item "pred": the predictions for the test set,
\item "thr.r2": the threshold used for clumping,
\item "set": the set of non-zero coefficients,
\item "timing": the execution time of the GWAS step only.
}
}
\description{
PRS (Clumping + Thresholding)
}
