% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-format-results.R
\name{read_format_results}
\alias{read_format_results}
\title{Read & format results}
\usage{
read_format_results(files, corr, ncores = nb_cores())
}
\arguments{
\item{files}{Files to read from.}

\item{corr}{Sparse correlation matrix of the genotype matrix.}

\item{ncores}{Number of cores to use. Default is \code{bigstatsr::nb_cores()}.}
}
\value{
A data frame of formated and aggregated results.
}
\description{
Read & format results
}
