% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ttrees}
\alias{ttrees}
\title{T-Trees}
\usage{
ttrees(TTree, file.base, pheno.all, ind.train, ind.test, n.trees = 100)
}
\arguments{
\item{TTree}{Path of TTree executable.}

\item{file.base}{Path of file containing the genotype matrix.}

\item{pheno.all}{All phenotypes.}

\item{ind.train}{Indices corresponding to the training set.}

\item{ind.test}{Indices corresponding to the test set.}

\item{n.trees}{Number of trees.}
}
\value{
A data frame with 5 variables:
\itemize{
\item "method": the name of the method,
\item "pred": the predictions for the test set,
\item "timing": the execution time,
\item "set": the set of non-zero coefficients.
}
}
\description{
T-Trees
}
